integrated ashenafi staff with the new entities
