+++ backend/src/common/entities/staff.entity.ts
@@ -1,23 +1,93 @@
-import { Column, Entity, OneToMany } from "typeorm";
-import { User } from "./user.entity";
-import { Assignment } from "./assignments.entity";
-
-@Entity()
-export class Staff extends User {
-
-    constructor()  {
-        super();
-    }
-
+// src/hms/staff/entities/staff.entity.ts
+import { 
+    Entity, 
+    PrimaryGeneratedColumn, 
+    Column, 
+    ManyToOne,
+    CreateDateColumn,
+    UpdateDateColumn,
+    BeforeInsert,
+    BeforeUpdate,
+    OneToMany
+  } from 'typeorm';
+  import * as bcrypt from 'bcrypt';
+  import { Hotel } from 'src/common/entities/hotel.entity';
+import { Assignment } from './assignments.entity';
+  
+  @Entity()
+  export class Staff {
+    @PrimaryGeneratedColumn('uuid')
+    id: string;
+  
     @Column()
-    salary: string;
-    
+    firstname: string;
+  
     @Column()
-    staffType: String;
+    lastname: string;
+  
+    @Column()
+    role: string;
+  
+    @Column({ unique: true })
+    email: string;
+  
+    @Column()
+    password: string;
+  
+    @Column({ default: true })
+    isTemporaryPassword: boolean;
+  
+    @Column({ 
+      type: 'enum',
+      enum: ['available', 'working'],
+      default: 'available'
+    })
+    status: string;
+  
+    @Column()
+    phonenumber: string;
+  
+    @Column({ nullable: true })
+    profilePic: string;
+  
+    @Column('decimal', { precision: 10, scale: 2 })
+    salary: number;
+  
+    // @Column({ type: 'date' })
+    // employedAt: Date;
+  
+    // @CreateDateColumn({ name: 'created_at' })
+    // createdAt: Date;
+  
+    // @UpdateDateColumn({ name: 'updated_at' })
+    // updatedAt: Date;
+  
+    @ManyToOne(() => Hotel, (hotel) => hotel.staff)
+    hotel: Hotel;
+  
+    @Column({ nullable: true })
+    currentTask: string;
+  
+    @Column({ nullable: true })
+    assignedRoomId: string;
 
-    @Column()
-    staffStatus: String;
 
     @OneToMany(() => Assignment, (assignment) => assignment.staff)
-    assignments: Assignment[];
-}
+    assignments: Assignment[]; //currently not being used
+    
+    @BeforeInsert()
+    @BeforeUpdate()
+    async hashPassword() {
+      if (this.password) {
+        this.password = await bcrypt.hash(this.password, 10);
+      }
+    }
+  
+    async validatePassword(password: string): Promise<boolean> {
+      return bcrypt.compare(password, this.password);
+    }
+  
+    get fullName(): string {
+      return `${this.firstname} ${this.lastname}`;
+    }
+  }
