+++ backend/src/v1/hms/staff/dto/assign-task.dto.ts
@@ -3,7 +3,6 @@ import { IsNotEmpty, IsString, IsUUID } from 'class-validator';
 
 export class AssignTaskDto {
   @IsNotEmpty()
-  @IsUUID()
   roomId: string;
 
   @IsNotEmpty()
