+++ backend/src/v1/hms/staff/staff.controller.ts
@@ -9,34 +9,29 @@ import {
   Headers, 
   UseGuards, 
   UsePipes,
-  ValidationPipe
+  ValidationPipe,
+  Query
 } from '@nestjs/common';
 import { StaffService } from './staff.service';
 import { CreateStaffDto } from './dto/create-staff.dto';
 import { AssignTaskDto } from './dto/assign-task.dto';
 import { GetStaffDto } from './dto/get-staff.dto';
-import { AuthGuard } from '../../auth/guards/auth.guard';
-import { RolesGuard } from '../../auth/guards/roles.guard';
-import { Roles } from '../../auth/decorators/roles.decorator';
-import { Role } from '../../auth/enums/role.enum';
 
-@Controller('api/v1/hms/staff')
-@UseGuards(AuthGuard, RolesGuard)
+@Controller('hms/staff')
 @UsePipes(new ValidationPipe())
 export class StaffController {
   constructor(private readonly staffService: StaffService) {}
 
   @Post()
-  @Roles(Role.Manager)
   async addStaff(
     @Headers('authorization') authorization: string,
     @Body() createStaffDto: CreateStaffDto,
   ) {
+    console.log(createStaffDto);
     return this.staffService.createStaff(createStaffDto);
   }
 
   @Patch(':id')
-  @Roles(Role.Manager)
   async assignStaffToRoom(
     @Headers('authorization') authorization: string,
     @Param('id') id: string,
@@ -45,9 +40,10 @@ export class StaffController {
     return this.staffService.assignTask(id, assignTaskDto);
   }
 
-  @Get()
-  @Roles(Role.Manager, Role.Admin)
-  async getAllStaff(@Body() getStaffDto: GetStaffDto) {
-    return this.staffService.getAllStaff(getStaffDto);
+  @Get('hotel/:hotelId')
+  async getAllStaff(
+    @Param('hotelId') hotelId: string,
+  ) {
+    return this.staffService.getAllStaff({hotelId});
   }
 }
